proto searches variables in the parent environment
the call to .$locale can return the locale function from readr package.
This occurs because proto looks the parent environment to search for variables.
For this reason I implemented a workaround checking if locale is a character.

look for better solutions, possibli a way to create a proto with an empty environment.