On tests copy files from extdata to tempdir
It is necessary copy files from extadata to tempdir when running tests for `read_marketdata`.

`read_marketdata` creates the parsed file in the same directory of the given file.
But files shouldn't be created on extdata, for this reason files must be copied to tempdir and used in the tests.

Created helper function copy_file_to_temp to copy files to tempdir.