dbdir and dataset_get
dbdir specifies the directory where the database files are stored. It is a db folder inside the cache folder. The default value is cachedir/db.

dataset_get returns an arrow dataset accoring to the given dataset name. At the first sight, each dataset refers to a template, but new datasets can be created from the template datasets.