new templates
template-test is to be used with the tests.
b3-bvbg-086 is used to the the scenario where one file has multiple files inside it. In this template the newer file is selected. It is done by sorting the files by its names . The file names have a timestamp in it. The file with the latest timestamp is selected.

The funtion select_file_if_multiple is used to select the file. It has a tag attribute that can be "newer" or "bigger". Indicating the newer file or the bigger file. The newer file is selected by sorting the files by its names. The bigger file is selected by sorting the files by its size. The default is NULL. If the tag is not set, the first file is selected.