Removed cachedir and clearcache functions
rb3_bootstrap handles the definitonand creation of rb3 cache folders.
The options rb3.cachedir defines the root directory for the cache.
Every time rb3.cachedir is set the function rb3_bootstrap must be executed to set the folders in the rb3_registry that is accessed by many function of the package.