b3-reference-rates brings different rate in columns depending on the curve name
Curve name PRE has 2 columns with rates - linear and compound.
Curve name DOC has 1 column and it is a linear rate.
Curve name DIC has 1 column and it is a compound rate.

For this reason the rate columns have generic names: col1 and col2.
In the specialized functions yc_*_get the column are selected and renamed correctly.